lib/requirements.txt
Flask==1.1.2
jsonschema==2.6.0
PyYAML==5.1.2
